move `_resolve_experiment` tests back to test_projects.py
